+++ cmn/src/main/java/core/aws/remote/RemoteResourceLoader.java
@@ -23,9 +23,9 @@ import core.aws.remote.vpc.SubnetLoader;
 import core.aws.remote.vpc.VPCLoader;
 import core.aws.resource.Resources;
 import core.aws.task.ec2.EC2TagHelper;
+import core.aws.util.Maps;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -36,14 +36,13 @@ public class RemoteResourceLoader {
     public void load(Environment env, Resources resources) {
         List<EnvTag> tags = loadEnvTags(env);
 
-        new SQSLoader(resources, env).load();
-        new S3Loader(resources, env).load();
-        new SNSLoader(resources, env).load();
-        new InstanceProfileLoader(resources, env).load();
+        new ImageLoader(resources, tags).load();
         new SGLoader(resources, tags).load();
         new KeyPairLoader(resources, env).load();
-        new InstanceLoader(resources, tags).load();
+        new InstanceLoader(resources, tags).load(); // instance loader must after sg, key pair, image loader to link remoteOnly sg/key and unfinished bake instance
+
         new VPCLoader(resources, tags).load();
+        new InstanceProfileLoader(resources, env).load();
         new NATLoader(resources, tags).load();
         new RouteTableLoader(resources, tags).load();
         new InternetGatewayLoader(resources, tags).load();
@@ -51,11 +50,14 @@ public class RemoteResourceLoader {
         new ServerCertLoader(resources, env).load();
         new ELBLoader(resources, env).load();
         new ASGroupLoader(resources, env).load();
-        new ImageLoader(resources, tags).load();  // TODO: image load needs to be after instance/key/sg loader for clean up, find better way?
+
+        new SQSLoader(resources, env).load();
+        new SNSLoader(resources, env).load();
+        new S3Loader(resources, env).load();
     }
 
     private List<EnvTag> loadEnvTags(Environment env) {
-        Map<String, EnvTag> tags = new HashMap<>();
+        Map<String, EnvTag> tags = Maps.newHashMap();
 
         DescribeTagsRequest request = new DescribeTagsRequest()
             .withFilters(new Filter("key").withValues(new EC2TagHelper(env).prefix() + ":*"));
@@ -63,11 +65,7 @@ public class RemoteResourceLoader {
 
         for (TagDescription remoteTag : remoteTags) {
             String remoteResourceId = remoteTag.getResourceId();
-            EnvTag tag = tags.get(remoteResourceId);
-            if (tag == null) {
-                tag = EnvTag.newTag(remoteTag);
-                tags.put(remoteResourceId, tag);
-            }
+            EnvTag tag = tags.computeIfAbsent(remoteResourceId, key -> new EnvTag(remoteTag));
             tag.addField(remoteTag);
         }