+++ cmn/src/main/java/core/aws/remote/ec2/ImageLoader.java
@@ -3,13 +3,10 @@ package core.aws.remote.ec2;
 import core.aws.remote.EnvTag;
 import core.aws.remote.Loader;
 import core.aws.resource.Resources;
-import core.aws.resource.ec2.Instance;
 import core.aws.resource.image.Image;
-import core.aws.util.StreamHelper;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
  * @author neo
@@ -30,17 +27,5 @@ public class ImageLoader extends Loader {
                 logger.warn("found unused image, name={}, version={}", tag.resourceId(), tag.version());
             }
         });
-
-        // link all unfinished ami instances
-        resources.stream()
-            .flatMap(StreamHelper.instanceOf(Image.class))
-            .forEach(image -> {
-                List<Instance> instances = resources.stream()
-                    .flatMap(StreamHelper.instanceOf(Instance.class))
-                    .filter(instance -> instance.id.startsWith("ami-" + image.id + "-"))
-                    .collect(Collectors.toList());
-
-                image.unfinishedBakeInstances.addAll(instances);
-            });
     }
 }